% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_management_df.R
\name{arrange_management_df}
\alias{arrange_management_df}
\title{arrange management_df by date, category}
\usage{
arrange_management_df(var_MGMT_data, include.combination = TRUE)
}
\arguments{
\item{var_MGMT_data}{management_df to be arranged}

\item{include.combination}{logical, if the combinations should considered
Default is TRUE}
}
\value{
a rearranged management_df
}
\description{
the funciton arranges the management_df by date and by category of operations.
 The order of the operations is harvest, fertilizer_application,crop_protection,
 tillage, sowing, irrigation, other
}
\examples{
#rearrange EXAMPLE data
arrange_management_df(EXAMPLE_data)

}
