% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indicators.R
\name{calculate_indicators}
\alias{calculate_indicators}
\title{Calculate all soil management indicators}
\usage{
calculate_indicators(var_MGMT_data)
}
\arguments{
\item{var_MGMT_data}{a \code{management_df} with a management history}
}
\value{
data frame with indices per year
}
\description{
Checks the \code{management_df} for consitency with the \code{\link[=check_management_df]{check_management_df()}} function.
Then it calculates the \code{\link[=C_input]{C_input()}}, \code{\link[=tillage_intensity]{tillage_intensity()}}, \code{\link[=soil_cover]{soil_cover()}},\code{\link[=plant_diversity]{plant_diversity()}}
\code{\link[=N_input]{N_input()}} and \code{\link[=productivity_indicator]{productivity_indicator()}}.
}
\examples{

\donttest{
  #example input
  calculate_indicators(EXAMPLE_data)
}

}
\seealso{
\itemize{
\item \code{\link[=check_management_df]{check_management_df()}}
\item \code{\link[=C_input]{C_input()}}
\item \code{\link[=tillage_intensity]{tillage_intensity()}}
\item \code{\link[=soil_cover]{soil_cover()}}
\item \code{\link[=N_input]{N_input()}}
\item \code{\link[=plant_diversity]{plant_diversity()}}
\item \code{\link[=productivity_indicator]{productivity_indicator()}}
}
}
