% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indicators_by_crop.R
\name{calculate_indicators_by_crop}
\alias{calculate_indicators_by_crop}
\title{Calculate soil management indicators by crop}
\usage{
calculate_indicators_by_crop(var_management_df)
}
\arguments{
\item{var_management_df}{a `management_df()`}
}
\value{
a tibble with the indicator values per crop
}
\description{
Alternative to `calculate_indicators()` that C inputs, N inputs, STIR and
soil cover for each crop separately.
The function identifies the last harvest of each crop per year and aggregates
the management intensities for the periods in between. Harvests of temporary 
leys that are sown after a main crop and harvest in the same year are not considered.
}
\examples{
# calculate examples
\donttest{
  calculate_indicators_by_crop(EXAMPLE_data)
}
# this would return a tibble, but it can take a while

}
