% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_N_input_tibble.R
\name{calculate_N_input_tibble}
\alias{calculate_N_input_tibble}
\title{Calculate N input tibble}
\usage{
calculate_N_input_tibble(var_MGMT_data)
}
\arguments{
\item{var_MGMT_data}{a \code{management_df} that contains the management information}
}
\value{
tibble with N and LSU values
}
\description{
A function that calculates the N input by mineral and organic fertilization
as well as the livestock intensity of the animal derived N inputs.
}
\details{
The function is mainly a helper function for the \code{\link[=N_input]{N_input()}} function.
}
\examples{
#Calculate N input tibble
calculate_N_input_tibble(EXAMPLE_data)
}
\seealso{
\itemize{
\item \code{\link[=calculate_indicators]{calculate_indicators()}} to calculate all management indicators
for a \code{management_df}
\item \code{\link[=CN_input_amendments]{CN_input_amendments()}} for the calculation of N inputs from organic amendments
\item \code{\link[=N_input]{N_input()}} a function that calculates the N input by mineral and organic fertilization and summarizes it by year
}
}
