% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pesticide_applications.R
\name{count_pesticide_applications}
\alias{count_pesticide_applications}
\title{Count number of pesticide applications}
\usage{
count_pesticide_applications(var_management_df)
}
\arguments{
\item{var_management_df}{management dataframe}
}
\value{
list with two tibbles: yearly and cropwise counts of pesticide applications
}
\description{
Counts the number of pesticide applications in the management dataframe. 
Besides, seperate counts of herbicide, fungicide and 
insecticide applications are provided.
}
\examples{
# count pesticide applications in the EXAMPLE_data
count_pesticide_applications(EXAMPLE_data)

}
