% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EXAMPLE_data}
\alias{EXAMPLE_data}
\title{Example of a management_df}
\format{
A tibble with 130 rows and 13 columns:
\describe{
  \item{crop}{Name of the main crop. Cover crop related operations are linked to the next main crop in the rotation [String from list]}
  \item{date}{Date of the management operation [Date]}
  \item{year}{Year of the management operation [Integer]}
  \item{category}{Categorization of the managment operation [1 level] [String from list]}
  \item{operation}{Categorization of the managment operation [2 level] [String from list]}
  \item{device}{Categorization of the managment operation [3 level] [String from list]}
  \item{value}{Numerical value linked to managment operation (e.g., depth of tillage operation, mass of organic amendment) [Integer]}
  \item{unit}{Unit of the numerical value (e.g. cm, t/ha) [String from list]}
  \item{machine}{Further information on the machine used (e.g., type, manufacturer, tool) [String [UTF-8]]}
  \item{product}{Further information on the applied product (e.g., name, manufacturer, C content) [String [UTF-8]]}
  \item{combination}{Indicate if a operation was done in combination with others.
    Use consequtive integer numbers if combinded operations occur.
    Leave empty if not combined}
  \item{comments}{Comments related to the management operation [String [UTF-8]]}
  \item{DMC}{Dry matter content of organic amendments [gDM/kgFM]}
  \item{C_content}{Carbon content of the amendments, relative to its dry matter [gC/kgDM]}
  \item{N_content}{Nitrogen content of organic amendments, relative to its dry matter [gN/kgDM]}
  \item{crop_product}{Crop product yield [tDM/ha]}
  \item{crop_residue}{Crop residue mass [tDM/ha]}
  \item{Cc_product}{Carbon content of the crop product [gC/kgDM$]}
  \item{Cc_residue}{Carbon content of the crop residue [gC/kgDM]}
}
}
\usage{
EXAMPLE_data
}
\description{
The dataset is derived from a Swiss long term agricultural field experiment.
 It is intended for demonstration purposes only
}
\keyword{datasets}
