% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_main_crops.R
\name{identify_main_crops}
\alias{identify_main_crops}
\title{Identify main crops}
\usage{
identify_main_crops(var_management_df)
}
\arguments{
\item{var_management_df}{management dataframe}
}
\value{
a management dataframe with a column "crop_no" that identifies the main crops
}
\description{
Identifies all main crops that are harvested in the management dataframe.
}
\examples{
# identify main crop in the EXAMPLE data
identify_main_crops(EXAMPLE_data)


}
