% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant_cover_LUT}
\alias{plant_cover_LUT}
\title{Look-up-table with default values to estimate soil cover by plants}
\format{
A tibble with 28 rows and 7 columns:
\describe{
  \item{Crop}{Name of the crop}
  \item{Slope_0_10}{Increase of crop cover per day between 0 and 10\% soil cover	[\%/day]}
  \item{Slope_10_50}{Increase of crop cover per day between 10 and 50\% soil cover	[\%/day]}
  \item{Slope_50_75}{Increase of crop cover per day between 50 and 75\% soil cover	[\%/day]}
  \item{Slope_75_100}{Increase of crop cover per day between 75 and 100\% soil cover	[\%/day]}
  \item{days_30}{Number of days it takes to reach 30\% soil cover	[day]}
  \item{Comments}{Source where the information was derived}
}
}
\usage{
plant_cover_LUT
}
\description{
The dataset is a look-up-table that is used to estimate the soil cover
 percentage by plant with the function `SoilManageR::plant_cover()`. 
 The data set is produced from the excel table `plant_cover_LUT.xlsx`
 file under `/inst/extdata/`.
}
\references{
\insertRef{mosimann2006}{SoilManageR}
}
\keyword{datasets}
