% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.soil_cover_tibble.R
\name{plot.soil_cover_tibble}
\alias{plot.soil_cover_tibble}
\title{Plotting soil_cover_tibbles}
\usage{
\method{plot}{soil_cover_tibble}(x, ...)
}
\arguments{
\item{x}{an soil_cover_tibble object}

\item{...}{arguments to be passed to methods}
}
\value{
none, it plots the soil_cover_tibble
}
\description{
This function plots objects of the class soil_cover_tibble in a custom format
}
\examples{
\donttest{
  data <- soil_cover(EXAMPLE_data, extended.output = TRUE)
  plot(data)
}


}
