% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilManageR-package.R
\docType{package}
\name{SoilManageR-package}
\alias{SoilManageR}
\alias{SoilManageR-package}
\title{SoilManageR: Calculate Soil Management Indicators for Agricultural Practice Assessment}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate numerical agricultural soil management indicators from on a management timeline of an arable field. Currently, indicators for carbon (C) input into the soil system, soil tillage intensity rating (STIR), number of soil cover and living plant cover days, N fertilization and livestock intensity, and plant diversity are implemented. The functions can also be used independently of the management timeline to calculate some indicators. The package contains tables with reference information for the functions, as well as a '*.xlsx' template to collect the management data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/SoilManageR/}
  \item Report bugs at \url{https://gitlab.com/SoilManageR/SoilManageR/-/issues}
}

}
\author{
\strong{Maintainer}: Olivier Heller \email{olivier.heller@agroscope.admin.ch} (\href{https://orcid.org/0000-0002-5918-4161}{ORCID})

Other contributors:
\itemize{
  \item Raphaël Wittwer \email{raphael.witter@agroscope.admin.ch} (\href{https://orcid.org/0000-0002-2129-7195}{ORCID}) [contributor]
}

}
\keyword{internal}
